requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
execnet==2.1.2
Faker==38.2.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
psycopg2==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.8.0
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.45
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
